requirements.txt
requests
pandas
python-dotenv
openai
streamlit